requirements.txt
# Core Framework
langgraph>=0.0.40
langchain>=0.1.4
langchain-community>=0.0.13
langchain-openai>=0.0.5

# Agent Communication (Fetch.ai)
uagents>=0.6.0
cosmpy>=0.9.0

# Vector Store
pinecone-client>=3.0.0
chromadb>=0.4.22

# LLM Inference
vllm>=0.2.7
openai>=1.10.0

# MCP Protocol
mcp>=0.1.0
httpx>=0.26.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
redis>=5.0.1
asyncpg>=0.29.0

# ML & Data
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.3

# Integrations
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0

# Monitoring
prometheus-client>=0.19.0
python-json-logger>=2.0.7

# x402 Payments (Coinbase stablecoin protocol)
# Note: x402 SDK when available, currently using httpx
web3>=6.0.0  # For wallet signing
eth-account>=0.10.0

# Security
pyjwt>=2.8.0
cryptography>=41.0.0
python-multipart>=0.0.6

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.1
mypy>=1.8.0
